Move QueryModalType ownership into the redux store
The redux store owned the uiControls.modal field but imported its type
from metabase/querying, an upward edge. The union now lives in
redux/store/qb.ts and MODAL_TYPES in querying/constants.ts is checked
against it with satisfies. All importers now take the type from
metabase/redux/store.
