--- frontend/src/metabase/redux/store/metabot.ts
@@ -1,1 +1,0 @@
-export { type MetabotState } from "metabase/metabot/state";
